requirements.txt
requests==2.31.0
python-dotenv==1.0.0
Flask==2.3.2
Werkzeug==3.0.1
Jinja2==3.1.2
MarkupSafe==2.1.3
click==8.1.7
itsdangerous==2.1.2
blinker==1.6.2
Flask-Caching==2.0.2
pytz==2023.3